feat: do actual typing check and some refactor
Add logic to use WordSelector to generate words in Toipe. These
generated words are then printed to the terminal in a faint style.

User's input is tracked and added to a vec of chars. Every char is
compared with the original text - if it matches the original, print in
normal style otherwise print in red underlined style.

Backspace is also handled - move back, print the original text in faint
style and last char is removed from the input vec.

Number of errors is also tracked and printed (although you can't see it
because screen is cleared xD)

Move initialization logic (reset screen, generate words) to new()
instead of a different function.