feat: calculate results and display them
Add a ToipeResults struct to store stats of the type test. Currenly only
stores number of chars, number of errors and start+end time of the test.
- Has methods to calculate duration of test, accuracy (0.0..=1.0) and
  characters per minute (cpm, speed)

After a test, results are displayed using a newly added
`display_results` function. Displays accuracy, number of errors and cpm
- This "window" waits for user input and if "r" key is pressed, the test
  is restarted.
- The important parts are highlighted in color.

The timer is started after the first key is pressed, but before
processing that key. For this purpose, the key processing function had
to be put in a new closure.

Also, the correctly typed text is now shown in green color because Faint
and Bold did not work reliably in some terminals (looking at you,
Konsole)