refactor: move multi word gen into WordSelector, store words
Toipe::gen_words intialized a new WordSelector everytime it was called.
This can get expensive as it has to read the entire wordlist everytime.
This commit will initialize a WordSelector in the ::new() fn and store
it in the Toipe struct for later use. The word generation will be done
in ::restart(). ::gen_words() will be renamed to ::show_words() since
it only does that now.

WordSelector::new_words will be added to make it easier to generate
multiple words.

The generated words vec will now stored inside the Toipe struct for
later use.