fix: prevent panic when there are words beyond Z/z
In some systems, the wordlist is sorted from A-Z and then a-z. The
current logic of the wordlist does not support this sorting.

This is a kludge that prevents panic by stopping when 26 letters have
been seen. This effectively makes it consider only the A-Z words on the
aforementioned systems, leading to a significantly smaller word list to
choose from.