feat: add a small 250 words list and use by default
Add a new `wordlists` module to select in-built wordlists. Only one word
list is currently provided, the `top250` list which includes the top 250
English words (by frequency). Source:
https://www.wordfrequency.info/samples.asp (top 60K lemmas sample)>
- The module also has the path to the OS wordlist which was previously
  the default.

The wordlist can be selected using the `-w` flag.

Refactor WordSelector to be general enough to handle any BufReader
(with Seek and Read implemented on the internal type). This allows
WordSelector to be made for a static string too.
- As a result of this, the Toipe struct now needs to store a boxed
  WordSelector since the implementation can vary based on whether a file
  or a string was selected.
- Separated the core logic into `RawWordSelector` which a generic struct
  for any BufReader (with constraints). It has specialized functions
  `from_path` and `from_string` to create appropriate `RawWordSelector`s
- Added a `WordSelector` trait that has `new_word` and `new_words`
  functions. This trait is implemented for all possible types of
  `WordSelector`.

Update `randomword` binary to use OS wordlist.

Update README with details on changing the wordlist.