fix: make accuracy, cpm and wpm more robust and accurate
You can't spam the keyboard to get a high score anymore :)

Accuracy is measured as percentage of correct chars in the total number
of chars typed (including mistakes that were erased using backspace).

cpm also uses number of correctly typed chars instead of all chars of
the original text.

wpm is now derived from cpm. This makes it only an approximation, but it
should be quite close. wpm is measured by first approximating the
average number of chars per word (using the number of characters and
words in the text).

Also renamed some fields to be less ambiguous (LSP rename is the best
thing).