feat: support navigating between lines
Completes most of #3

This required tracking of the position of all lines (of words) and the
cursor position - which line and char the user is currently on.

Added a `LinePos` struct:
- stores position and length of a line

Added a `CursorPos` struct:
- stores the positions of all lines
- stores the current line and the character in that line the user is on
- has methods to go to the next/prev lines
- maps line no./char no. pair to the x-y position on screen (terminal)

Added helper functions in `ToipeTui`:
- to move the cursor front/back
- to move the cursor to current position (in the lines)
- to get the current line index

Modified `replace_text` to use the above helpers. This makes it only
work for single character inputs - clarified in doc string.

Line tracking is only done when a `ToipeTui.track_lines` is set. Need to
find a better way to enable this in certain contexts only - perhaps
requires another overhaul of the code.

This also introduces a `reset` method in `ToipeTui` since `CursorPos` is
stateful. It's called in `display_words` to avoid wrong usage (without
calling `reset`).