feat: document keyboard shortcuts
Display keyboard shortcuts at the bottom of the typing test screen. Also
moved shortcuts to the bottom on the results screen.
- Added a helper function `display_lines_bottom` for displaying stuff
  at the bottom of the screen (currently used only to display
  shortcuts).
- only two shortcuts are displayed here.

Documented all keyboard shortcuts in the about section of the CLI help
page. Also added a section for this in the README.

Also modified Cargo.toml's description - removed "tux"