Handle all cases #25
added handling for the 3 possible cases that RawWordSelector creation
can fail in Toipe::new(). However, there is still code duplication in
the match statemen. I haven't figured out a way around this because the
compiler does not shee the result returned from
RawWordSelector::from_path() and RawWordSelector::from_string() as the
same type